+++ django/utils/asyncio.py
@@ -1,6 +1,8 @@
 import asyncio
 import functools
+import inspect
 
+from asgiref.sync import sync_to_async
 from django.core.exceptions import SynchronousOnlyOperation
 
 
@@ -30,3 +32,43 @@ def async_unsafe(message):
         return decorator(func)
     else:
         return decorator
+
+
+def auto_async(func):
+    """
+    Decorator to automatically convert a sync function to async depending on
+    the calling context.
+    """
+
+    def sync_wrapper(*args, **kwargs):
+        return func(*args, **kwargs)
+
+    async def async_wrapper(*args, **kwargs):
+        return await sync_to_async(func)(*args, **kwargs)
+
+    def inner(*args, **kwargs):
+        # Initial experiment with frame hacks.
+        # This needs to be expended for other possible use cases.
+        #
+        # This may also break easily. What happens if the parent wants a
+        # coroutine to later pass it around?
+        try:
+            outer_frames = inspect.getouterframes(inspect.currentframe())
+        except IndexError:
+            # I'm not sure if this only happens on ipython, but for now
+            # forcing async here helps me debug
+            # ToDo: Review this.
+            return async_wrapper(*args, **kwargs)
+        try:
+            parent_frame = outer_frames[1]
+            if parent_frame.frame.f_code.co_flags & (
+                inspect.CO_COROUTINE |
+                inspect.CO_ITERABLE_COROUTINE |
+                inspect.CO_ASYNC_GENERATOR):
+                return async_wrapper(*args, **kwargs)
+        except IndexError:
+            pass  # No outer frames
+
+        return sync_wrapper(*args, **kwargs)
+
+    return inner
